requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asttokens==2.4.1
audioop-lts==0.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.1.0
Flask==3.1.0
fonttools==4.55.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==0.21.0
google-cloud-language==2.15.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
jsonpickle==4.0.0
kiwisolver==1.4.7
langdetect==1.0.9
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.1.3
oauth2client==3.0.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
pipwin==0.5.2
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyjsparser==2.7.1
pyparsing==3.2.0
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pyvis==0.3.2
pywin32==308
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.12.0
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
gunicorn
